FingerPrint_Sensor_R307/requirements.txt
pyfingerprint>=1.5
RPi.GPIO>=0.7.0; platform_machine=="armv7l" or platform_machine=="aarch64"
